ça marche presque

+++ directory_manager_thread.py
@@ -1,19 +1,32 @@
 import Directory
 import os
 import threading
+from logger import Logger
+
+import time
+
 
 def thingModified(self, listAllThingModified, lock):
+
     while listAllThingModified:
+
         with lock:
-            elem = listAllThingModified.pop()
-        if elem[0] == "file":
-            if elem[1] == "deleteAndCreate":
-                self.ftp.remove_file(elem[3])
-                self.ftp.file_transfer(elem[2], elem[3], elem[4])
-            elif elem[1] == "create":
-                self.ftp.file_transfer(elem[2], elem[3], elem[4])
-        else:
-            if elem[1] == "created":
-                self.ftp.create_folder(elem[2])
+            try:
+                elem = listAllThingModified.pop()
+            except:
+                break
+        print(len(listAllThingModified))
 
+        if elem:
+            self.ftp.connect()
+            if elem[0] == "file":
+                if elem[1] == "deleteAndCreate":
+                    self.ftp.remove_file(elem[3])
+                    self.ftp.file_transfer(*elem[2:])
+                elif elem[1] == "create":
+                    self.ftp.file_transfer(*elem[2:])
+            else:
+                if elem[1] == "created":
+                    self.ftp.create_folder(elem[2])
 
+            self.ftp.disconnect()
